+++ matcher/preset.ts
@@ -1,5 +1,8 @@
 // Copyright 2021-Present the Unitest authors. All rights reserved. MIT license.
 // This module is browser compatible.
+import { toHaveBeenCalledOnce } from "./to_have_been_called_once.ts";
+import { toHaveBeenCalledAfter } from "./to_have_been_called_after.ts";
+import { toHaveBeenCalledBefore } from "./to_have_been_called_before.ts";
 import { toContainEqual } from "./to_contain_equal.ts";
 import { toIncludeSameMembers } from "./to_include_same_members.ts";
 import { toIncludeAnyMembers } from "./to_include_any_members.ts";
@@ -132,6 +135,9 @@ const jestMatcherMap = {
  * @see https://github.com/jest-community/jest-extended
  */
 const jestExtendedMatcherMap = {
+  toHaveBeenCalledOnce,
+  toHaveBeenCalledAfter,
+  toHaveBeenCalledBefore,
   toIncludeSameMembers,
   toIncludeAnyMembers,
   toBeEmpty,