+++ mock/utils_test.ts
@@ -1,5 +1,10 @@
 // Copyright 2021-Present the Unitest authors. All rights reserved. MIT license.
-import { isReturn, isTypeReturn, pickValue } from "./utils.ts";
+import {
+  incrementalNumber,
+  isReturn,
+  isTypeReturn,
+  pickValue,
+} from "./utils.ts";
 import { assertEquals } from "../dev_deps.ts";
 
 Deno.test("isReturn", () => {
@@ -13,4 +18,13 @@ Deno.test("isTypeReturn", () => {
 Deno.test("pickValue", () => {
   assertEquals(pickValue({ type: "return", value: 0 }), 0);
   assertEquals(pickValue({ type: "return", value: 1 }), 1);
+});
+
+Deno.test({
+  name: "incrementalNumber",
+  fn: () => {
+    assertEquals(incrementalNumber(), 1);
+    assertEquals(incrementalNumber(), 2);
+    assertEquals(incrementalNumber(), 3);
+  },
 });